Break FetchingPromptsPage out separately
gh-395

So I can test history-related interactions, including with WebSockets.
